Updated Readme, included Mac drivers, added new packages, setup pre-commit checks
