+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "FrogGames",
+  "name": "frog_god_games_m3",
   "version": "1.0.0",
   "main": "index.js",
   "repository": "https://github.com/triscuitcircuit/Moonwake-DnD.git",
@@ -10,21 +10,41 @@
     "@shopify/shopify-api": "^5.1.0",
     "@vitejs/plugin-react": "^2.1.0",
     "ngrok": "^4.3.3",
+    "oracledb": "file:drivers/mac-arm/oracledb-5.5.0.tgz",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "sequelize": "^6.28.0",
     "vite": "^3.1.4"
   },
   "devDependencies": {
-    "@types/node": "^18.7.23",
+    "@types/node": "^18.11.18",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.48.2",
+    "@typescript-eslint/parser": "^5.48.2",
     "dotenv": "^16.0.3",
-    "typescript": "^4.8.4"
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-standard-with-typescript": "^30.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^15.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.32.1",
+    "husky": "^8.0.0",
+    "lint-staged": "^13.1.0",
+    "prettier": "^2.8.3",
+    "typescript": "^4.9.4"
   },
   "scripts": {
     "build": "npx tsc && vite build",
     "prestart": "yarn run build",
     "start": "node dist/index.js",
-    "preview": "npx tsc && vite build && vite preview"
+    "preview": "npx tsc && vite build && vite preview",
+    "prepare": "husky install",
+    "pre-commit": "lint-staged"
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx,html,css}": "npx prettier --check"
   }
 }
