--- src/Box.ts
@@ -1,5 +1,0 @@
-import { styled } from "@nextui-org/react"
-
-export const Box = styled("div", {
-    boxSizing: "border-box",
-});
