+++ src/Display/Box.ts
@@ -1,0 +1,5 @@
+import { styled } from "@nextui-org/react";
+
+export const Box = styled("div", {
+  boxSizing: "border-box",
+});
