--- src/FrogGodLogo.js
@@ -1,11 +1,0 @@
-export const FrogGodLogo = () => (
-    <svg
-        className=""
-        fill="none"
-        height="36"
-        viewBox="0 0 32 32"
-        width="36"
-        xmlns="http://www.w3.org/2000/svg"
-    >
-    </svg>
-);
